backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
channels==4.0.0
channels-redis==4.1.0
psycopg2-binary==2.9.9
redis==5.0.1
Pillow==10.1.0
django-filter==23.5
django-ckeditor==6.7.0
celery==5.3.4
django-celery-beat==2.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.29.7
reportlab==4.0.7
openpyxl==3.1.2
django-extensions==3.2.3

